refactor(button) - extract types to a dedicated file
