+++ src/components/UI/Button/Button.tsx
@@ -2,23 +2,13 @@
 
 import classNames from 'classnames'
 
-import styles from './Button.module.scss'
+import { ButtonProps, ButtonType } from '@/types/button.types'
 
-export enum ButtonType {
-  Primary = 'primary',
-  Secondary = 'secondary',
-}
+import styles from './Button.module.scss'
 
 const buttonClasses: Record<ButtonType, string> = {
   [ButtonType.Primary]: styles.primary,
   [ButtonType.Secondary]: styles.secondary,
-}
-
-export interface ButtonProps {
-  children: React.ReactNode
-  handleClick: (e: React.MouseEvent<HTMLButtonElement>) => void
-  type?: ButtonType
-  className?: string
 }
 
 const Button = ({
