+++ src/components/UI/index.ts
@@ -5,6 +5,5 @@ export { default as Text } from './Text/Text'
 export { TextTag, TextType } from './Text/Text'
 
 export { default as Button } from './Button/Button'
-export { ButtonType, type ButtonProps } from './Button/Button'
 
 export { default as List } from './List/List'
